+++ src/vector.h
@@ -4,33 +4,33 @@
 class Vector
 {
 public:
-	double x;
-	double y;
-	double z;
-	
+	float x;
+	float y;
+	float z;
+
 	Vector();
-	Vector(double _x, double _y, double _z);
+	Vector(float _x, float _y, float _z);
 	Vector(const Vector& rhs);
 
-	double length() const;
-	double lengthsq() const;
-	double dotprod(const Vector& rhs) const;
+	float length() const;
+	float lengthsq() const;
+	float dotprod(const Vector& rhs) const;
 	Vector crossprod(const Vector& rhs) const;
-	
+
 	void normalise();
-	
+
 	void rotateAroundX(double angle);
 	void rotateAroundY(double angle);
-	
+
 	Vector operator+(const Vector& rhs) const;
 	Vector operator-(const Vector& rhs) const;
 	Vector& operator+=(const Vector& rhs);
 	Vector& operator-=(const Vector& rhs);
-	
-	Vector operator*(double f) const;
-	Vector operator/(double f) const;
-	Vector& operator*=(double f);
-	Vector& operator/=(double f);
+
+	Vector operator*(float f) const;
+	Vector operator/(float f) const;
+	Vector& operator*=(float f);
+	Vector& operator/=(float f);
 	
 	bool operator==(const Vector& rhs) const;
 	bool operator!=(const Vector& rhs) const;
